[Update: 방향 전환 1프레임 지연, Prev Dir 개념 추가 및 방향결정구조 재구성] [Fix: 방향감시 사용하는 State들 수정] [Update: Dash State]

+++ Projects/Engine/CAnim.cpp
@@ -44,7 +44,7 @@ void CAnim::finaltick()
 
 		if (m_CurFrmIdx >= m_vecFrm.size())
 		{
-			m_CurFrmIdx = m_vecFrm.size() - 1;
+			m_CurFrmIdx = (int)m_vecFrm.size() - 1;
 			m_bFinish = true;
 		}
 		m_AccTime = 0.f;
